Fixed namespaces, moved queryableperson to gateway folder
