+++ HousingSearchApi.Tests/V1/Helper/SearchPersonsQueryContainerOrchestratorTests.cs
@@ -1,7 +1,7 @@
 using System.Linq;
 using FluentAssertions;
 using HousingSearchApi.V1.Boundary.Requests;
-using HousingSearchApi.V1.Factories;
+using HousingSearchApi.V1.Gateways;
 using HousingSearchApi.V1.Interfaces;
 using Nest;
 using Xunit;
