+++ HousingSearchApi.Tests/V1/Helper/TestDataHelper.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using AutoFixture;
-using HousingSearchApi.V1.Factories;
+using HousingSearchApi.V1.Gateways;
 using Nest;
 
 namespace HousingSearchApi.Tests.V1.Helper
