+++ HousingSearchApi/V1/Interfaces/ISearchPersonQueryContainer.cs
@@ -1,5 +1,5 @@
 using HousingSearchApi.V1.Boundary.Requests;
-using HousingSearchApi.V1.Factories;
+using HousingSearchApi.V1.Gateways;
 using Nest;
 
 namespace HousingSearchApi.V1.Interfaces
