+++ HousingSearchApi/V1/Interfaces/SearchPersonElasticSearchHelper.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Amazon.Lambda.Core;
 using HousingSearchApi.V1.Boundary.Requests;
-using HousingSearchApi.V1.Factories;
+using HousingSearchApi.V1.Gateways;
 using HousingSearchApi.V1.Interfaces.Sorting;
 using Nest;
